requirements.txt
# Core
pyspark==3.5.0
findspark==2.0.1
pandas==2.2.0
python-dotenv==1.0.1

# Orchestration
apache-airflow==2.9.3

# DB clients
psycopg2-binary==2.9.9
sqlalchemy==2.0.27

# Azure SDK (production)
azure-storage-blob==12.23.0

# Utilities
python-dateutil==2.9.0